Modified main execution
Checking for arguments has been moved to the executable and render_files takes
the source and destination as arguments. These default to '.' and './build' for
the d executable.

Moving argument checking to the executable and passing source and destination into
d.render_files allows custom executables, like the included 'bin/da' example.